cohort class

This class create a cohort object, which will hold many information related to a cohort including cohort ID, name, description, filters, tables. Use of this object will simplify calling every other functions which requires a cohort. This can be created using constructor function cb_create_cohort or cb_load_cohort

Slots

id

cohort ID.

name

cohort name.

desc

cohort description.

fields

Filters.

more_fields

Filter related information.